requirements.txt
cryptography==3.4.7
pycrypto==3.10.1